Сделал задание 5.1
